--- airbnb/src/main/java/com/airbnb/airbnb/entity/Category.java
@@ -1,30 +1,0 @@
-package com.airbnb.airbnb.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
-
-@Getter
-@Setter
-@Entity
-@Table
-@AllArgsConstructor
-@NoArgsConstructor
-public class Category {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(nullable = false)
-    private Long id;
-
-    @Column(nullable = false)
-    private String categoryName;
-
-    @OneToMany(mappedBy = "category")
-    private List<StayCategories> stayCategories = new ArrayList<>();
-}
