--- airbnb/src/main/java/com/airbnb/airbnb/entity/StayCategories.java
@@ -1,38 +1,0 @@
-package com.airbnb.airbnb.entity;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.*;
-
-@Getter
-@Setter
-@Entity
-@Table
-@AllArgsConstructor
-@NoArgsConstructor
-public class StayCategories {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "stay_id")
-    @JsonIgnore
-    private Stay stay;
-
-    @ManyToOne
-    @JoinColumn(name = "category_id")
-    @JsonIgnore
-    private Category category;
-
-    public StayCategories(Stay stay, Category verifiedCategory) {
-        this.stay = stay;
-        this.category = verifiedCategory;
-    }
-
-}
