+++ airbnb/src/main/java/com/airbnb/airbnb/image/entity/Image.java
@@ -1,0 +1,39 @@
+package com.airbnb.airbnb.image.entity;
+
+import com.airbnb.airbnb.stay.entity.Stay;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+@Entity
+@Table
+@AllArgsConstructor
+@NoArgsConstructor
+public class Image {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(nullable = false)
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name = "stay_id")
+    @JsonIgnore
+    private Stay stay;
+
+    @Column(nullable = false)
+    private String fileName;
+
+    @Column(nullable = false)
+    private String imageUrl;
+
+    @Column(nullable = false)
+    private LocalDateTime uploadedAt;
+}
