+++ airbnb/src/main/java/com/airbnb/airbnb/member/dto/MemberPostDto.java
@@ -1,6 +1,5 @@
-package com.airbnb.airbnb.dto;
+package com.airbnb.airbnb.member.dto;
 
-import com.airbnb.airbnb.entity.UserRoleEnum;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
