+++ airbnb/src/main/java/com/airbnb/airbnb/repository/ImageRepository.java
@@ -1,6 +1,6 @@
 package com.airbnb.airbnb.repository;
 
-import com.airbnb.airbnb.entity.Image;
+import com.airbnb.airbnb.image.entity.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ImageRepository extends JpaRepository<Image,Long> {
