--- airbnb/src/main/java/com/airbnb/airbnb/service/StayCategoriesService.java
@@ -1,11 +1,0 @@
-package com.airbnb.airbnb.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@RequiredArgsConstructor
-@Service
-public class StayCategoriesService {
-
-
-}
