+++ airbnb/src/main/java/com/airbnb/airbnb/stay/service/StayCategoriesService.java
@@ -1,0 +1,11 @@
+package com.airbnb.airbnb.stay.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@RequiredArgsConstructor
+@Service
+public class StayCategoriesService {
+
+
+}
